Added support for the photo manager's secondary screen, which displays the full-sized image corresponding to the thumbnail selected on the main screen. This screen consists of two tabs: one to show the full-sized version of the selected thumbnail and another to show a map of the vicinity of where the picture was acquired.
The picture tab has a menu that allows users to either play back the
audio tag associated with the diplayed image or to record a new tag for
it. The two menu commands are just stubs for now.

Furthermore, nothing is actually displayed yet in either of the two tabs.
At this point, the whole thing is just skeletal infrastructure...