Removed the whole set-up for trying to pass back the thumbnail ID of the newly captured image from the ImageRecorder to the main activity. This is either impossible or really quite convoluted. The difficulty arises from the fact that Android is running and updating the camera device, the thumbnails database and the activity stack all in separate threads. Thus, when the ImageRecorder activity finishes and returns control to the main activity, the camera has not yet called ImageCaptureCallback.onPictureTaken(). Consequently, the main activity will always see the returned thumbnail ID as -1.
In the mean time, when the ImageCaptureCallback.onPictureTaken() function
finally does get called, it won't get the correct thumbnail ID either
because the database update happens in yet another thread...

Coordinating between all these threads (that are internal to Android) is
not a straightforward thing to implement. So we forget about
communicating thumbnail IDs between ImageRecorder and the main activity.
What this means is that the user will be returned to the thumbnails
gallery after taking a new picture instead of going directly to the
application's display screen. While that feature would have been nice,
being directed to the thumbnails gallery as an intermediat step is only a
minor inconvenience.