Added preliminary support for displaying the map of the selected picture's acquisition vicinity. An informal test using fake location data shows that the feature works. However, the Google Maps view is showing only a grid and ddms constantly shows HTTP errors... Hopefully, this is not something completely intractable.
The assignment is nearly complete. We just have to test everything on the
phone and make sure all the features work. Hopefully, the camera
preview-and-capture will magically fix itself and start working again;
hopefully, the audio stuff will work out of the box; and, hopefully, so
will the map tab.

Of course, we still have to add that pin-on-the-map feature for showing
exactly where the selected picture was taken. Let's hope this is not too
convoluted. And let's not forget that ImageRecorder.java still needs the
correct API calls for getting a GPS fix on the phone (whatever version of
the SDK it's using)...