Added the ability to go directly from the camera preview to the display screen without having to first return to the thumbnails page. It turns out that the earlier notions about Android's multiple threads for the Camera device, media database and activity stack (see svn log for r17 of AssignmentThree.java), while correct, were not entirely germane to the issue at hand.
The trick is to use an intent to directly start the display screen from
the camera preview. As for the fact that the display screen expects the
selected image's thumbnail ID to be passed to it as part of the extras
bundle: all we had to do was look up the thumbnail ID of the newly
created image.

Wish this realization had come a couple of days earlier! Oh well, live
and learn...