Added support for deleting images and their corresponding audio tags.
There is some general, unfathomable weirdness about the application. For
the most part, it works fine; but every now and then, it crashes. Don't
know why.

Tried adding a means of removing all pictures and audio tags acquired by
the application in one shot. But that was a disaster! The code for it is
commented out for now (AssignmentThree.java).