Gave the MapTab code, which was produced during amateur hour, a complete professional make-over.
This is pretty much the last check-in for this assignment. There is still
one very elusive crash to fix. Nonetheless, this is the version I wish
we'd submitted... Anyway, hopefully, the knowledge gained hacking this
code-base the past couple of days will help in the fourth assignment and
for the project...